requirements.txt
streamlit
pandas
websocket-client
numpy
watchdog
